Fixed two bugs in selectTopic and sendMessage
- selectTopic
currentOperation->tmp is used as temporary variable to send the message,
but the data structure is being freed two times from two different
pointers in the operationDestroy.
This can happen if the client is removed after sending the message (from
case 1 of selectTopic)
The messageArray is destroyed at the destroying of the sendMessage from
the p variable but nothing resets the tmp variable on the previous
operation, so on the previous operation, the tmp variable is being
destroyed twice (as it should contains a received message)

- sendMessage
After sending the message the server passes directly to the recvCommand,
this makes the server irresponsive in case the client is irresponsive
too, until the client disconnects.
Instead the Operation needs to be called from the select after the
that some data is available